+++ stdlib/public/runtime/Private.h
@@ -530,7 +530,7 @@ public:
     
     Demangle::NodePointer operator()(Demangle::SymbolicReferenceKind kind,
                                      Demangle::Directness isIndirect,
-                                     int32_t offset,
+                                     int64_t offset,
                                      const void *base);
   };
